DEV: Split weekly job into multiple smaller jobs
The weekly job can take more than 2 hours to run on larger sites.
It is ideal for the jobs to be as small as possible and this is what
this commit attempts.
