+++ app/jobs/scheduled/clean_up_drafts.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Jobs
+  class CleanUpDrafts < ::Jobs::Scheduled
+    every 1.week
+
+    def execute(args)
+      Draft.cleanup!
+    end
+  end
+end
