+++ app/jobs/scheduled/clean_up_user_auth_tokens.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Jobs
+  class CleanUpUserAuthTokens < ::Jobs::Scheduled
+    every 1.week
+
+    def execute(args)
+      UserAuthToken.cleanup!
+    end
+  end
+end
